requirements.txt
frida~=16.0.19
Pillow~=9.5.0
psutil~=5.9.5
numpy~=1.24.3